pyproject.toml
[tool.poetry]
name = "gpu_exporter"
version = "0.0.1"
description = "Collect data about GPUs and exposes them for Prometheus"
authors = ["Carlos Rincon <c.rincon@addermobile.com>"]

[tool.poetry.dependencies]
python = "^3.7"
requests = "^2.28.0"
prometheus-client = "^0.14.1"
toml = "^0.10.2"
PyEventEmitter = "^1.0.5"
sty = "^1.0.4"
python-dotenv = "^0.20.0"
click = "^8.1.3"

[tool.poetry.dev-dependencies]
black = "^22.3.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
